+++ model/labels/labels_stringlabels.go
@@ -286,8 +286,8 @@ func (ls Labels) WithoutEmpty() Labels {
 // ByteSize returns the approximate size of the labels in bytes.
 // String header size is ignored because it should be amortized to zero
 // because it may be shared across multiple copies of the Labels.
-func (ls Labels) ByteSize() int {
-	return len(ls.data)
+func (ls Labels) ByteSize() uint64 {
+	return uint64(len(ls.data))
 }
 
 // Equal returns whether the two label sets are equal.
